requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose[cryptography]
pytest
pytest-asyncio
